pyproject.toml
[project]
name = "e-commerke"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "flake8>=7.3.0",
    "isort>=6.1.0",
    "load-dotenv>=0.1.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "python3-distutils>=0.0.1",
]
